error processing added

+++ src/main/java/com/easy/app/user/UserService.java
@@ -1,5 +1,7 @@
 package com.easy.app.user;
 
+import com.easy.app.user.exception.BadRequestException;
+import com.easy.app.user.exception.UserNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,12 +18,18 @@ class UserService {
     }
 
     void addUser(User user) {
-        //todo impl validation
+        final Boolean existByEmail = userRepository.isExistByEmail(user.getEmail());
+        if (existByEmail) {
+            throw new BadRequestException(String.format("User with email %s already exists", user.getEmail()));
+        }
         userRepository.save(user);
     }
 
     void deleteUser(long userId) {
-        //todo check if user exist
+        final boolean userExist = userRepository.existsById(userId);
+        if (!userExist) {
+            throw new UserNotFoundException(String.format("User not found by id={%s}", userId));
+        }
         userRepository.deleteById(userId);
     }
 }
